Add files via upload
Name and surname field should be set to be capitalised e.g. from “john” to
“John” before being inserted into DB -- Done
The script should validate the email address before inserting to make sure that
it is valid (valid means that it is a legal email format e.g. “xxxx@asdf@asdf is
not a legal format). In the instance that an email is invalid, no insert should be
made to database and error message reported to STDOUT -- WIP